Health checks for helidon SE.
