+++ examples/microprofile/hello-world-explicit/pom.xml
@@ -40,7 +40,7 @@
     <dependencies>
         <dependency>
             <groupId>io.helidon.microprofile.bundles</groupId>
-            <artifactId>helidon-microprofile-1.1</artifactId>
+            <artifactId>helidon-microprofile-1.2</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
