+++ health/health-checks/src/main/resources/META-INF/beans.xml
