+++ health/health/src/main/java/io/helidon/health/HealthCheckException.java
@@ -13,18 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package io.helidon.health;
 
 /**
- * Aggregator module for microprofile 1.2.
+ * Exception thrown by Health system if something goes wrong.
  */
-module io.helidon.microprofile.v1_2 {
-    requires transitive io.helidon.mp.config.cdi;
-    requires transitive io.helidon.mp.config;
-    requires transitive io.helidon.mp.server;
-    requires transitive io.helidon.mp.health;
-    requires transitive io.helidon.mp.metrics;
-    requires transitive io.helidon.mp.faulttolerance;
-    requires transitive io.helidon.mp.jwt.auth.cdi;
-
-    requires io.helidon.health.checks;
+public class HealthCheckException extends RuntimeException {
+    /**
+     * Exception with a message and a cause.
+     *
+     * @param message descriptive message
+     * @param cause   cause of this exception
+     */
+    public HealthCheckException(String message, Throwable cause) {
+        super(message, cause);
+    }
 }
