+++ microprofile/bundles/helidon-microprofile-1.2/pom.xml
@@ -37,12 +37,12 @@
         </dependency>
         <dependency>
             <groupId>io.helidon.microprofile.health</groupId>
-            <artifactId>helidon-microprofile-health-core</artifactId>
+            <artifactId>helidon-microprofile-health</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>io.helidon.microprofile.health</groupId>
-            <artifactId>helidon-microprofile-health-checks</artifactId>
+            <groupId>io.helidon.health</groupId>
+            <artifactId>helidon-health-checks</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
