--- microprofile/health/health-core/src/test/java/io/helidon/microprofile/health/HealthCheckResponseImplTest.java
@@ -1,51 +1,0 @@
-/*
- * Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.helidon.microprofile.health;
-
-import java.util.HashMap;
-
-import org.eclipse.microprofile.health.HealthCheckResponse;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-class HealthCheckResponseImplTest {
-
-    @Test
-    void nullNameThrowsException() {
-        assertThrows(NullPointerException.class, () -> new HealthCheckResponseImpl(
-                null,
-                HealthCheckResponse.State.DOWN,
-                new HashMap<>()));
-    }
-
-    @Test
-    void nullStateThrowsException() {
-        assertThrows(NullPointerException.class, () -> new HealthCheckResponseImpl(
-                "Bad",
-                null,
-                new HashMap<>()));
-    }
-
-    @Test
-    void nullDataThrowsException() {
-        assertThrows(NullPointerException.class, () -> new HealthCheckResponseImpl(
-                "Bad",
-                HealthCheckResponse.State.DOWN,
-                null));
-    }
-}
