+++ microprofile/health/src/main/resources/META-INF/beans.xml
