+++ microprofile/tests/tck/tck-health/pom.xml
@@ -39,7 +39,7 @@
         </dependency>
         <dependency>
             <groupId>io.helidon.microprofile.health</groupId>
-            <artifactId>helidon-microprofile-health-core</artifactId>
+            <artifactId>helidon-microprofile-health</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
